created utils module
